basis: epoch-tagged agent streams + idempotent delete RPCs
Two correctness fixes plus a small consolidation of shared state:

1. Stale agent stream cleanup no longer clobbers a faster reconnect.
   Each `stream_messages` call mints a process-unique `epoch`; the
   per-stream cleanup runs `agents.remove_if(host, |a| a.epoch == mine)`
   so an old stream that lost the race only releases its own slot.
   Pending CreateMachine waiters and the `agent_connected` metric
   ride the same compare-and-remove, so a slow disconnect handler
   can no longer cancel in-flight ops or zero metrics belonging to
   the live connection.

2. `DeleteMachine` and `DeleteCluster` are now idempotent at the
   server boundary. Once the final tombstone ack drops the row, a
   retried CAPI finalizer cleanup observed NotFound forever and
   stranded the Kubernetes object. Treating NotFound as success
   makes the contract honestly idempotent for every caller.

Drive-bys:
- Move `pending_ops` onto `SharedCtx`; it was being clone-passed
  into both services and threaded through `handle_agent_message`
  for no benefit.
- Extract the disconnect cleanup as a free `release_agent_stream`
  with explicit deps so the race fix has direct unit coverage
  without needing a full SharedCtx.
- Drop a stray `as u32` no-op cast in db.rs that newer clippy
  flagged.

+++ crates/basis-controller/src/db.rs
@@ -2218,7 +2218,7 @@ mod tests {
         assert_eq!(u32::from(n.bridge_start), u32::from(n.cidr.network()) + 1);
         assert_eq!(
             u32::from(n.bridge_end) - u32::from(n.bridge_start),
-            (net.bridge_reserve - 1) as u32
+            net.bridge_reserve - 1
         );
         assert_eq!(u32::from(n.vm_start), u32::from(n.bridge_end) + 1);
         assert_eq!(u32::from(n.vm_end), u32::from(n.cidr.broadcast()) - 1);
@@ -2811,7 +2811,10 @@ mod tests {
     #[tokio::test]
     async fn update_cluster_trust_domain_overwrites_stored_value() {
         let db = test_db().await;
-        let net = db.allocate_cluster_network(&make_net_config()).await.unwrap();
+        let net = db
+            .allocate_cluster_network(&make_net_config())
+            .await
+            .unwrap();
         let mut cluster = make_cluster("c1", "c1", net, "10.100.0.1");
         cluster.trust_domain = "old-td".to_string();
         db.insert_cluster(&cluster).await.unwrap();
